# src/main/resources/openapi/pennywise-v1.yaml
openapi: 3.0.3
info:
  title: PennyWise API
  version: 1.0.0
  description: Personal expense tracking API with multi-book support, idempotent operations, and optimistic concurrency control
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /v1/books:
    get:
      tags: [Books]
      summary: List books for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/BookResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Books]
      summary: Create book (currency locked)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookCreateRequest" }
      responses:
        "201":
          description: Created
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/books/{bookId}:
    parameters:
      - $ref: "#/components/parameters/BookId"
    get:
      tags: [Books]
      summary: Get book
      responses:
        "200":
          description: OK
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Books]
      summary: Update book (name only; currency locked)
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookUpdateRequest" }
      responses:
        "200":
          description: OK
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
...
  /v1/books/{bookId}/budgets:
    parameters:
      - $ref: "#/components/parameters/BookId"
      - name: month
        in: query
        required: true
        schema:
          type: string
          pattern: '^\d{4}-\d{2}$'
          example: "2026-01"
    get:
      tags: [Budgets]
      summary: List monthly category budgets with spend
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/BudgetResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/books/{bookId}/budgets/{categoryId}:
    parameters:
      - $ref: "#/components/parameters/BookId"
      - $ref: "#/components/parameters/CategoryId"
      - name: month
        in: query
        required: true
        schema:
          type: string
          pattern: '^\d{4}-\d{2}$'
          example: "2026-01"
    put:
      tags: [Budgets]
      summary: Create or update a monthly category budget
      parameters:
        - $ref: "#/components/parameters/OptionalIfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BudgetUpsertRequest" }
      responses:
        "200":
          description: OK
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BudgetResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
    delete:
      tags: [Budgets]
      summary: Remove a monthly category budget
      parameters:
        - $ref: "#/components/parameters/OptionalIfMatch"
      responses:
        "204":
          description: Deleted
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        example: '"3"'
      description: ETag value from prior GET (optimistic concurrency)
    OptionalIfMatch:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
        example: '"3"'
      description: Optional ETag value for optimistic concurrency when updating an existing resource
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    PreconditionFailed:
      description: Precondition Failed (ETag mismatch)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string, example: "Invalid request" }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            requestId: { type: string, example: "01HZX..." }
      required: [error]
...
    BookCreateRequest:
      type: object
      required: [name, currencyCode, timezone]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        currencyCode: { type: string, minLength: 3, maxLength: 3, example: "USD" }
        timezone: { type: string, example: "America/Chicago" }
        openingBalanceMinor: { type: integer, format: int64, example: 0 }
    BookUpdateRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
    BookResponse:
      type: object
      required:
        [id, name, currencyCode, timezone, openingBalanceMinor, createdAt, updatedAt, version]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        currencyCode: { type: string }
        timezone: { type: string }
        openingBalanceMinor: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        version: { type: integer, format: int64 }
...
    BudgetUpsertRequest:
      type: object
      required: [amountMinor]
      properties:
        amountMinor: { type: integer, format: int64, minimum: 0 }
    BudgetResponse:
      type: object
      required:
        [
          id,
          bookId,
          categoryId,
          categoryName,
          month,
          amountMinor,
          spentMinor,
          remainingMinor,
          currencyCode,
          createdAt,
          updatedAt,
          version
        ]
      properties:
        id: { type: string, format: uuid }
        bookId: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        categoryName: { type: string }
        month: { type: string, example: "2026-01" }
        amountMinor: { type: integer, format: int64 }
        spentMinor: { type: integer, format: int64 }
        remainingMinor: { type: integer, format: int64 }
        currencyCode: { type: string, example: "USD" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, format: int64 }
...